Mucking up my codebase cause python displays weird behaviour.
For reference: The logic included in test class A mimics the logic with TimeSeriesData and _index_dimension. Yet the latter fails in another project to access the property and instead skips to __getattr__. Yet I cannot recreate this behavior directly inside the project, suggesting that it only happens because the code is located in an external module.